Implements default layout from previous portfolio project.
Left to do:

- [ ] Implement menu
- [ ] Make queries to the TakeShape API
