+++ shape-portfolio-gatsbyjs/src/components/footer.js
@@ -1,0 +1,9 @@
+import React from "react"
+
+const Footer = () => (
+  <footer className="footer">
+    <p>© {new Date().getFullYear()} <a href="https://takeshape.io" target="_blank">TakeShape Inc.</a></p>
+  </footer>
+)
+
+export default Footer
