+++ shape-portfolio-gatsbyjs/src/components/menu.js
@@ -1,0 +1,7 @@
+import React from "react"
+
+const Menu = () => (
+  <div></div>
+)
+
+export default Menu
