+++ shape-portfolio-gatsbyjs/src/pages/404.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Layout from '../components/layout'
+import Layout from '../layouts/default'
 
 const NotFoundPage = () => (
   <Layout>
