+++ shape-portfolio-gatsbyjs/src/pages/page-2.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
-import Layout from '../components/layout'
+import Layout from '../layouts/default'
 
 const SecondPage = () => (
   <Layout>
